Adding permissions to Clearchat.cs

+++ BeetleBot/Modules/ClearChat.cs
@@ -13,19 +13,24 @@ namespace BeetleBot.Modules
         [Command("clearchat")]
         public async Task ClearChatAsync(int clearAmount)
         {
-            if (clearAmount > 0 && clearAmount < 101)
+            if (Program.hasPermissions(Context.Guild.GetUser(Context.User.Id)))
             {
-                //==============================Delete Command Message=========================
-                await Context.Message.DeleteAsync();
-                //=============================================================================
-                
-                var messages = await Context.Channel.GetMessagesAsync(clearAmount).FlattenAsync();
-                var reversed = messages.Reverse();
-                foreach (IMessage msg in reversed)
-                    Program.SendLog(Context.User, "ClearChat", msg);
-                await ((ITextChannel)Context.Channel).DeleteMessagesAsync(messages);
+                if (clearAmount > 0 && clearAmount < 101)
+                {
+                    //==============================Delete Command Message=========================
+                    await Context.Message.DeleteAsync();
+                    //=============================================================================
+
+                    var messages = await Context.Channel.GetMessagesAsync(clearAmount).FlattenAsync();
+                    var reversed = messages.Reverse();
+                    foreach (IMessage msg in reversed)
+                        Program.SendLog(Context.User, "ClearChat", msg);
+                    await ((ITextChannel)Context.Channel).DeleteMessagesAsync(messages);
+                }
             }
-            
+            else
+                await ReplyAsync(Context.User.Mention + ", you do not have permission to clear chat.");
+
         }
     }
 }
